requirements.txt
openai>=1.0.0
pandas>=1.3.0
faiss-cpu>=1.7.0 
pickle-mixin>=1.0.2 